Check for commonly misspelled words during CI build
Use codespell to check the repository's files for typos.

In the event of false positives, words to ignore may be added to etc/codespell-ignore-words-list.txt. Although the typo check is case-insensitive, the ignore words list is case-sensitive, and the misspelled words list is all lower case, so always use lower case for the ignore words list.